Renaming stuff and separating title screen from game.
