+++ engine/src/core/engine_context.h
@@ -2,8 +2,8 @@
 
 #include <stdbool.h>
 
-#define DEFAULT_START_PROJECT_PATH "test_games/cardboard_fighter"
-//#define DEFAULT_START_PROJECT_PATH "test_games/card_test"
+//#define DEFAULT_START_PROJECT_PATH "test_games/cardboard_fighter"
+#define DEFAULT_START_PROJECT_PATH "test_games/card_test"
 
 typedef struct CREEngineContext {
     bool isRunning;
